+++ packages/core/src/connectivity/scp-cf/xsuaa-service.ts
@@ -172,6 +172,7 @@ export async function jwtBearerTokenGrant(
 }
 
 /**
+ * Deprecated since v1.49.0. Use `fetchVerificationKeys(url: string)` instead. Credentials are ignored.
  * Fetches verification keys from the XSUAA service for the given credentials.
  * @param xsuaaCredentials - Credentials of the XSUAA service instance.
  * @param jku - Value of the jku property in the JWT header. If not provided the old legacy URL xsuaaCredentials.url/token_keys is used as a fallback which will not work for subscriber accounts created after 14th of April 2020.
@@ -183,11 +184,12 @@ export function fetchVerificationKeys(
 ): Promise<TokenKey[]>;
 
 /**
+ * Deprecated since v1.49.0. Use `fetchVerificationKeys(url: string)` instead. Credentials are ignored.
  * Fetches verification keys from the XSUAA service for the given URL, with the given pair of credentials.
- * @param URL - URL of the XSUAA service instance.
+ * @param url - URL of the XSUAA service instance.
  * @param clientId - Client ID of the XSUAA service instance.
  * @param clientSecret - Client secret of the XSUAA service instance.
- * @returns An array of TokenKeys.
+ * @returns An array of token keys.
  */
 export function fetchVerificationKeys(
   url: string,
@@ -195,9 +197,17 @@ export function fetchVerificationKeys(
   clientSecret: string
 ): Promise<TokenKey[]>;
 
+/**
+ * Fetches verification keys from the XSUAA service for the given URL.
+ * @param url - URL of the XSUAA service instance.
+ * @returns An array of token keys.
+ */
+export function fetchVerificationKeys(url: string): Promise<TokenKey[]>;
+
 export function fetchVerificationKeys(
   xsuaaUriOrCredentials: string | XsuaaServiceCredentials,
   clientIdOrJku?: string,
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   clientSecret?: string
 ): Promise<TokenKey[]> {
   // The case where the XsuaaServiceCredentials are given as object
@@ -206,34 +216,32 @@ export function fetchVerificationKeys(
       logger.warn(
         'JKU field from the JWT not provided. Use xsuaaClient.url/token_keys as fallback. ' +
           'This will not work for subscriber accounts created after 14th of April 2020.' +
-          'Please provide the right URL given by the field JKU present in the JWT header.'
+          'Please provide the right URL given by the field `jku` in the JWT header.'
+      );
+      return executeFetchVerificationKeys(
+        `${xsuaaUriOrCredentials.url}/token_keys`
       );
     }
-    return fetchVerificationKeys(
-      clientIdOrJku || `${xsuaaUriOrCredentials.url}/token_keys`,
-      xsuaaUriOrCredentials.clientid,
-      xsuaaUriOrCredentials.clientsecret
-    );
+
+    return executeFetchVerificationKeys(clientIdOrJku);
   }
   // The three strings case
+  return executeFetchVerificationKeys(xsuaaUriOrCredentials);
+}
+
+// TODO: in v2 move this implementation to `fetchVerificationKeys`
+function executeFetchVerificationKeys(url: string): Promise<TokenKey[]> {
   const config: AxiosRequestConfig = {
-    url: xsuaaUriOrCredentials,
+    url,
     method: 'GET'
   };
-  if (clientIdOrJku && clientSecret) {
-    const authHeader = headerForClientCredentials({
-      username: clientIdOrJku,
-      password: clientSecret
-    });
-    config.headers = { Authorization: authHeader };
-  }
 
   return axios
     .request(config)
     .then(resp => resp.data.keys.map(k => renameKeys(tokenKeyKeyMapping, k)))
     .catch(error => {
       throw new ErrorWithCause(
-        `Failed to fetch verification keys from XSUAA service instance ${xsuaaUriOrCredentials}!`,
+        `Failed to fetch verification keys from XSUAA service instance "${url}".`,
         error
       );
     });